requirements.txt
polars==1.12.0
SQLAlchemy==2.0.36
fastapi==0.115.4
python-dotenv==1.0.1
imagekitio==4.1.0
psycopg2_binary==2.9.10
sqlmodel==0.0.22
gspread==6.1.4
oauth2client==4.1.3